int64_t OpenCL device count (#969)
* Converting OpenCL device IDs to 64-bit

* Debug 64-bit device ID instances

+++ include/qengine_cpu.hpp
@@ -46,8 +46,8 @@ protected:
 public:
     QEngineCPU(bitLenInt qBitCount, bitCapInt initState, qrack_rand_gen_ptr rgp = nullptr,
         complex phaseFac = CMPLX_DEFAULT_ARG, bool doNorm = false, bool randomGlobalPhase = true, bool ignored = false,
-        int ignored2 = -1, bool useHardwareRNG = true, bool useSparseStateVec = false,
-        real1_f norm_thresh = REAL1_EPSILON, std::vector<int> ignored3 = {}, bitLenInt ignored4 = 0U,
+        int64_t ignored2 = -1, bool useHardwareRNG = true, bool useSparseStateVec = false,
+        real1_f norm_thresh = REAL1_EPSILON, std::vector<int64_t> ignored3 = {}, bitLenInt ignored4 = 0U,
         real1_f ignored5 = FP_NORM_EPSILON_F);
 
     ~QEngineCPU() { Dump(); }
@@ -75,7 +75,7 @@ public:
 #endif
     }
 
-    void SetDevice(int dID) {}
+    void SetDevice(int64_t dID) {}
 
     real1_f FirstNonzeroPhase()
     {
