+++ include/qinterface.hpp
@@ -2246,7 +2246,7 @@ public:
     /**
      *  Set the device index, if more than one device is available.
      */
-    virtual void SetDevice(int dID) = 0;
+    virtual void SetDevice(int64_t dID) = 0;
 
     /**
      *  Get the device index. ("-1" is default).
